Moved code from ProductCard index.ts to HomePage index.ts and then deleted the ts file in ProductCard. Changed the imports, code works like it should
